FriendlyDoLoop.jl

FriendlyDoLoop.jl

Dummy documentation pages created for testing purposes

FriendlyDoLoop.CellPointToSValue — Type.

Abstract type representing a collection of values defined on a set of local points in each of the cells of the mesh. if cell_point_to_value is an instance of this type value = cell_point_to_value[cell][point] will be the value associated with the cell id cell, and local point id point

source